Use c2nim to read msp430 header.
The ed script to make the header usable is horrific.
It should be ported to nim itself, at which point I will need
a filtered sort of include (translating line by line into
statements with a macro?). Basically TI's abuse of the C
preprocessor is too much for c2nim to handle.